update code for the golangci-lint release v.2.7.0
Remove some of the gosec type linter check exceptions. Latest version of
golanci-lint can determine the bounds of variables better and does not
raise problems.

Split the check_os_updates code into available platforms of: linux,
darwin, windows, other. Fix the linting problems that would arise due to
unused variables on different platforms.
