+++ pkg/humanize/humanize.go
@@ -122,7 +122,7 @@ func NumF(num int64, precision int) string {
 		prefix = "-"
 	}
 
-	return prefix + humanizeBytes(uint64(num), 1000, []string{"", "k", "M", "G", "T", "P", "E"}, precision) //nolint:gosec // underflow checked before but not recognized by gosec
+	return prefix + humanizeBytes(uint64(num), 1000, []string{"", "k", "M", "G", "T", "P", "E"}, precision)
 }
 
 // Bytes(82854982) -> 83 MB
